mandelbrot_alpha

+++ includes/core/core_point.h
@@ -12,27 +12,27 @@
 
 typedef struct	s_point3
 {
-	float		x;
-	float		y;
-	float		z;
+	int			x;
+	int			y;
+	int			z;
 }				t_point3;
 
 typedef struct	s_point2
 {
-	float		x;
-	float		y;
+	int			x;
+	int			y;
 }				t_point2;
 
 typedef struct	s_point2c
 {
-	float		x;
-	float		y;
+	int			x;
+	int			y;
 	t_color		c;
 }				t_point2c;
 
-t_point3		*p_point3(float x, float y, float z);
-t_point3		point3(float x, float y, float z);
-t_point2		*p_point2(float x, float y);
-t_point2		point2(float x, float y);
-t_point2c		point2c(float x, float y, t_color c);
+t_point3		*p_point3(int x, int y, int z);
+t_point3		point3(int x, int y, int z);
+t_point2		*p_point2(int x, int y);
+t_point2		point2(int x, int y);
+t_point2c		point2c(int x, int y, t_color c);
 #endif
