+++ src/core/core_point.c
@@ -7,7 +7,7 @@
 
 #include "core/core_point.h"
 
-t_point3		*p_point3(float x, float y, float z)
+t_point3		*p_point3(int x, int y, int z)
 {
 	t_point3	*point;
 	
@@ -19,12 +19,12 @@ t_point3		*p_point3(float x, float y, float z)
 	return (point);
 }
 
-t_point3		point3(float x, float y, float z)
+t_point3		point3(int x, int y, int z)
 {
 	return (t_point3){ x, y, z };
 }
 
-t_point2		*p_point2(float x, float y)
+t_point2		*p_point2(int x, int y)
 {
 	t_point2	*point;
 	
@@ -35,12 +35,12 @@ t_point2		*p_point2(float x, float y)
 	return (point);
 }
 
-t_point2		point2(float x, float y)
+t_point2		point2(int x, int y)
 {
 	return (t_point2){ x, y };
 }
 
-t_point2c		point2c(float x, float y, t_color c)
+t_point2c		point2c(int x, int y, t_color c)
 {
 	return (t_point2c){ x, y, c };
 }
